sns http support - minor refactoring and added helper methods
