+++ sns/http_notifications_test.go
@@ -1,5 +1,11 @@
 package sns_test
 
+import (
+	"encoding/json"
+	"github.com/crowdmob/goamz/sns"
+	"gopkg.in/check.v1"
+)
+
 var HttpSubscriptionConfirmationRequest = `
 {
   "Type" : "SubscriptionConfirmation",
@@ -44,3 +50,36 @@ var HttpUnsubscribeConfirmationRequest = `
   "SigningCertURL" : "https://sns.us-east-1.amazonaws.com/SimpleNotificationService-f3ecfb7224c7233fe7bb5f59f96de52f.pem"
 }
 `
+
+func (s *S) TestHttpSubscriptionConfirmationRequestUnmarshalling(c *check.C) {
+	notification := sns.HttpNotification{}
+	err := json.Unmarshal([]byte(HttpSubscriptionConfirmationRequest), &notification)
+	c.Assert(err, check.IsNil)
+	c.Assert(notification.Token, check.NotNil)
+	c.Assert(notification.Subject, check.Equals, "")
+	c.Assert(notification.SubscribeURL, check.NotNil)
+	c.Assert(notification.UnsubscribeURL, check.Equals, "")
+	c.Assert(notification.Type, check.Equals, sns.MESSAGE_TYPE_SUBSCRIPTION_CONFIRMATION)
+}
+
+func (s *S) TestHttpNotificationRequestUnmarshalling(c *check.C) {
+	notification := sns.HttpNotification{}
+	err := json.Unmarshal([]byte(HttpNotificationRequest), &notification)
+	c.Assert(err, check.IsNil)
+	c.Assert(notification.Token, check.Equals, "")
+	c.Assert(notification.Subject, check.NotNil)
+	c.Assert(notification.SubscribeURL, check.Equals, "")
+	c.Assert(notification.UnsubscribeURL, check.NotNil)
+	c.Assert(notification.Type, check.Equals, sns.MESSAGE_TYPE_NOTIFICATION)
+}
+
+func (s *S) TestHttpUnsubscribeConfirmationRequestUnmarshalling(c *check.C) {
+	notification := sns.HttpNotification{}
+	err := json.Unmarshal([]byte(HttpUnsubscribeConfirmationRequest), &notification)
+	c.Assert(err, check.IsNil)
+	c.Assert(notification.Token, check.NotNil)
+	c.Assert(notification.Subject, check.Equals, "")
+	c.Assert(notification.SubscribeURL, check.NotNil)
+	c.Assert(notification.UnsubscribeURL, check.Equals, "")
+	c.Assert(notification.Type, check.Equals, sns.MESSAGE_TYPE_UNSUBSCRIBE_CONFIRMATION)
+}
